requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
antlr4-python3-runtime==4.9.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cn2an==0.5.19
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
demucs==3.0.6
diffq==0.2.3
dora-search==0.1.11
einops==0.6.0
eng-to-ipa==0.0.2
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.92.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.9.0
Flask==2.3.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
gradio==3.19.1
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
indic-transliteration==2.3.37
inflect==6.0.2
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
julius==0.2.7
kiwisolver==1.4.4
ko-pron==1.3
lameenc==1.4.2
lazy_loader==0.1
librosa==0.10.0
linkify-it-py==2.0.0
llvmlite==0.39.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
networkx==3.1
nltk==3.4.5
num-thai==0.0.5
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.4
openai-whisper==20230124
OpenCC==1.1.1
openpyxl==3.0.10
openunmix==1.2.1
orjson==3.8.6
packaging==23.0
pandas==1.5.2
pefile==2022.5.30
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
pooch==1.7.0
proces==0.1.3
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
pydub==0.25.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyopenjtalk==0.2.0
pyparsing==3.0.9
pypinyin==0.48.0
PyQt5==5.15.8
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.6
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3986==1.5.0
roman==3.3
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.5
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.3
SpeechRecognition==3.9.0
starlette==0.25.0
submitit==1.4.5
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.8.2+cu110
tqdm==4.64.1
transformers==4.26.1
treetable==0.2.5
typer==0.7.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
uc-micro-py==1.0.1
Unidecode==1.3.6
urllib3==1.23
uvicorn==0.20.0
websockets==10.4
Werkzeug==2.3.3
yarl==1.8.2
zipp==3.15.0
